+++ InternalsViewer.Internals/Interfaces/Services/Loaders/IAllocationChainService.cs
@@ -7,7 +7,7 @@ namespace InternalsViewer.Internals.Interfaces.Services.Loaders;
 
 public interface IAllocationChainService
 {
-    Task<AllocationChain> LoadChain(Database database, int fileId, PageType pageType);
+    Task<AllocationChain> LoadChain(Database database, short fileId, PageType pageType);
 
     Task<AllocationChain> LoadChain(Database database, PageAddress startPageAddress);
 }
