+++ InternalsViewer.Internals/Interfaces/Services/Loaders/IPfsChainService.cs
@@ -5,5 +5,5 @@ namespace InternalsViewer.Internals.Interfaces.Services.Loaders;
 
 public interface IPfsChainService
 {
-    Task<PfsChain> LoadChain(Database database, int fileId);
+    Task<PfsChain> LoadChain(Database database, short fileId);
 }
